moving table hook upper for better control over table states
